+++ app/backend/lib/emails/templates/notifyDocumentUpload.ts
@@ -3,13 +3,30 @@ import {
   EmailTemplateProvider,
 } from '../handleEmailNotification';
 
+const getUploadedFileList = (
+  documentNames: string[],
+  documentTypes: string[]
+) =>
+  documentNames
+    .map((file, index) => {
+      const documentType =
+        documentTypes.length === 1 ? documentTypes[0] : documentTypes[index];
+
+      return `<li><em>${file}</em>${documentType ? ` (${documentType})` : ''}</li>`;
+    })
+    .join('');
+
 const notifyDocumentUpload: EmailTemplateProvider = (
   applicationId: string,
   url: string,
   initiator: any,
   params: any
 ): EmailTemplate => {
-  const { ccbcNumber, documentType, timestamp, documentNames } = params;
+  const { ccbcNumber, documentTypes, timestamp, documentNames, rfiNumber } =
+    params;
+
+  let documentTypeFormatted =
+    documentTypes.length > 1 ? 'Multiple files' : `A ${documentTypes[0]}`;
 
   const section = {
     'Claim & Progress Report': 'project?section=claimsReport',
@@ -17,19 +34,27 @@ const notifyDocumentUpload: EmailTemplateProvider = (
     'Milestone Report': 'project?section=milestoneReport',
     'Statement of Work': 'project?section=projectInformation',
   };
+  let link = `<a href='${url}/analyst/application/${applicationId}/${section[documentTypes[0]] ?? 'rfi'}'>${ccbcNumber}</a>`;
 
-  const link = `<a href='${url}/analyst/application/${applicationId}/${section[documentType] ?? 'rfi'}'>${ccbcNumber}</a>`;
+  if (rfiNumber) {
+    link = `<a href='${url}/analyst/application/${applicationId}/rfi'>RFI ${rfiNumber}</a>`;
+    documentTypeFormatted = documentTypes.includes('Email Correspondence')
+      ? 'RFI Email Correspondence'
+      : documentTypeFormatted;
+  }
+
   return {
     emailTo: [112, 10, 111],
     emailCC: [],
     tag: 'document-upload-notification',
-    subject: `${documentType} uploaded in Portal`,
+    subject: `${documentTypeFormatted} uploaded in Portal`,
     body: `
-        <h1>${documentType} uploaded in Portal</h1>
+        <h1>${documentTypeFormatted} uploaded in Portal</h1>
 
-        <p>Notification: A ${documentType} has been uploaded in the Portal for ${link} on ${timestamp}.</p>
+        <p>Notification: ${documentTypeFormatted} has been uploaded in the Portal for ${link} on ${timestamp}.</p>
+        <p>File(s) uploaded:</p>
         <ul>
-          ${documentNames.map((file) => `<li><em>${file}</em></li>`).join('')}
+          ${getUploadedFileList(documentNames, documentTypes)}
         </ul>
     `,
   };